+++ packages/flutter/lib/mountain_view.dart
@@ -1,0 +1,11 @@
+// Copyright 2017 The Chromium Authors. All rights reserved.
+// Use of this source code is governed by a BSD-style license that can be
+// found in the LICENSE file.
+
+/// Flutter widgets implementing base Android UI components.
+///
+/// To use, import `package:flutter/mountain_view.dart`.
+
+library mountain_view;
+
+export 'src/mountain_view/page.dart';
